requirements.txt
langchain 
langchain-community 
langchain-google-genai 
python-dotenv
pydantic
duckduckgo-search
beautifulsoup4
psycopg2-binary
django 
djangorestframework